+++ Runtime/Scripts/Extensions/Integrals/Short/ShortExtensions.Sqrt.cs
@@ -6,9 +6,9 @@ namespace NumericMath
 
 	public static partial class ShortExtensions
 	{
-		public static float Sqrt(this short value)
+		public static double Sqrt(this short value)
 		{
-			return (float)Math.Sqrt(value);
+			return Math.Sqrt(value);
 		}
 	}
 }
